Fix #8 Be more specific about what image to return
Instead of returning the entire Dockerfile, detect cases where the
range is between two images and consider those ranges to be a part of
the image that it succeeds and not what it precedes.

The argument here is that if there are comments before the
declaration of a new build stage then presumably those comments are
for the build stage that follows and not the build stage that just
finished. This assumption is in place given that Dockerfiles are
written linearly from top to bottom so any comments in the file
presumably apply to what comes after instead of what came before.

Signed-off-by: Remy Suen <remy.suen@gmail.com>