Add an isDefined() function to Variable
Add a new isDefined() API to Variable to allow clients to determine
whether the variable has been defined or not. A variable is
considered to be defined if an ARG or ENV instruction prior to it has
defined a variable with the same name. Variables that refer to one of
the Docker builder's default variables are considered to be
undefined.

Signed-off-by: Remy Suen <remy.suen@gmail.com>