Fix #24 Handle spaces when replacing variables
If spaces are used within a quoted string for a variable replacement,
the parser was incorrectly considering the quoted string as two
separate arguments instead of a single argument. If a quoted string
is detected within a variable, spaces should be ignored and
considered a part of the containing argument instead of as an
argument delimiter.

Signed-off-by: Remy Suen <remy.suen@gmail.com>