Fix #25 Rename getRawValue() to getUnescapedValue()
The name getRawValue() is misleading as it carries an implication
that it includes all the content contained within the range of the
value. However, escaped newline characters and embedded comments are
not included in this function's return value so the function has been
renamed to a slightly more accurate getUnescapedValue().

Signed-off-by: Remy Suen <remy.suen@gmail.com>