Fix #37 Make getPropertyArguments() public
As PropertyInstruction's getArguments() has been changed to be the
same implementation as Instruction, callers of the originally
overridden function may no longer retrieve the same thing in some
scenarios. To alleviate this, getPropertyArguments() is now public so
that callers of the original getArguments() can change their code to
call getPropertyArguments() instead to get the same behaviour as
before.

Signed-off-by: Remy Suen <remy.suen@gmail.com>