Fix #54 Append comments to build stages
Only Dockerfiles were returning comments if asked via getComments().
Now build stages will also return comments if getComments() is
invoked on them.

Note that only comments that are contained within the FROM
instruction and the last instruction within the build stage will be
included. Comments that come before a FROM instruction are not
considered to be a comment of that build stage.

Dockerfile's getContainingImage() has been changed to consider the
case above. Thus, requesting the containing image of a position
within a comment before a build stage will return the Dockerfile
itself instead of the build stage.

Signed-off-by: Remy Suen <remy.suen@gmail.com>