Fix #53 Add getRange() function to ImageTemplate
Created a new convenience function for retrieving the range of the
Dockerfile or build stage. Note that the range of a build stage only
includes the first FROM instruction and the last instruction before
the following FROM instruction. Any comments between the final
instruction and the FROM instruction of the next build stage will not
be considered to be in the range of either build stage.

Signed-off-by: Remy Suen <remy.suen@gmail.com>