feat: tsc 3: alias ImageTemplate
typescript 3 (including in client code) isn't happy with the duplicate use
of ImageTemplate as a class and as an interface here. I believe the intention
is to expose the interface, so maintain that behaviour, and instead alias
the class as we import it and extend it, and do not expose it.

Giving the two types different names, and exporting them, would probably be
clearer, but would be a breaking change?