Merge conditional checks for \r with \n
We have a lot of code that takes \r but skips ahead by 1 to jump over
the following \n. While this is a nice optimization, it does make the
code more complicated than it needs to be and it is unlikely that this
optimization is the deciding factor in whether we parse quickly or not.
Readability and maintainability is more important to us than parsing
speed for the foreseeable future so we can merge the conditional checks
together.

Signed-off-by: Remy Suen <remy.suen@gmail.com>