Use more functions instead of one big one
The parse(*) function was over 200 lines long and involved many nested
iteration loops (of which many were also named loops) with various
different kinds of checks for breaking out early if necessary. As the
Dockerfile specification evolves, the function will only grow larger so
it has been refactored to improve readadability and maintainability.

Signed-off-by: Remy Suen <remy.suen@gmail.com>